qcow2: Use block-queue
This changes qcow2 to use block-queue for metadata I/O, so that the number of
bdrv_flush calls during heavy cluster allocation is greatly reduced.

Most of this is mechanical conversion of bdrv_pwrite_sync to a sequence of
blkqueue_pwrite and blkqueue_barrier (or bdrv_pread -> blkqueue_pread). Also,
most functions get a QcowRequest structure now instead of a BlockDriverState.
The QcowRequest contains the current block-queue context and refers to the
BlockDriverState of the image.

The more interesting parts include the bdrv_(aio_)flush implementation which
has to deal with errors.

Signed-off-by: Kevin Wolf <kwolf@redhat.com>