coroutine: Fix nested coroutine calls
Consider the following scenario:

1. A enters B
2. B enters C
3. C yields, control returns to B
4. B terminates

The correct action to take is that control returns to A. This is implemented by
this patch.

In the old code, control would return to C because B was entered by a context
switch from C. At the same time 'current' would be inconsistent with this and
point to A; additionally C->caller is NULL. This leads very quickly to a crash.

Signed-off-by: Kevin Wolf <kwolf@redhat.com>