block: remove unnecessary code; image always r/w if snapshot=on
Important note: this patch should only be applied after patch
 "block: do not allow read-only=on and snapshot=on to be used together"

Now that 'read-only=on' and 'snapshot=on' are not allowed together,
forcing open_flags to have BDRV_O_RDWR set is redundant code, and can be
removed.  This is because unless read-only=on is specified, the image
will always be opened with BDRV_O_RDWR set.

Signed-off-by: Jeff Cody <jcody@redhat.com>
Reviewed-by: Eric Blake <eblake@redhat.com>
Signed-off-by: Kevin Wolf <kwolf@redhat.com>