qapi: Add "backing" and "file" to BlockStats
Currently there is no way to query BlockStats of the backing chain. This
adds "backing" and "file" fields into BlockStats to make it possible.

The old optional field "parent", which has the information for "file",
is now an alias for backward compatibility. This gets a better naming
consistency with "file" option name in the block device configuration
interface.

Signed-off-by: Fam Zheng <famz@redhat.com>