qemu-img: Improve error messages
Previously, when there is a user error in argv parsing, qemu-img prints
help text and exits.

Add an error_exit function to print a helpful error message and a hint
to run 'qemu-img --help' for more information.

As a bonus, "qemu-img <cmd> --help" now has a more reasonable exit code
0.

In the future the help text should be split by sub command, and only
print the information for the specified command.

Signed-off-by: Fam Zheng <famz@redhat.com>

---
v3: Remove "return 0" in the end of main. (Sen)
v2: Address Eric's comments on QEMU_NORETURN, article and
    EXIT_{SUCCEESS,FAILURE}.