qapi: Allow setting default values for optional parameters
In command definition, 'defaults' is now parsed as a dict of default
values. Only optional parameters will have effect in generated code.

'str' and 'int' are supported. In generated code, 'str' will be
converted to g_strdup'ed string pointer, 'int' will be identically
assigned.

"has_foo" will be skipped in decalarations as if it's a required
argument if default value is specified in the schema.

String default values are g_strdup'ed so the dealloc visitor can g_free
on it correctly.

E.g.

    { 'command': 'block-commit',
      'data': { 'device': 'str', '*base': 'str', 'top': 'str',
                '*speed': 'int' },
      'defaults': {'base': 'earthquake', 'speed': 100 } }

will generate

    int qmp_marshal_input_block_commit(Monitor *mon, const QDict *qdict, QObject **ret)
    {
        ...
        char * base = g_strdup("earthquake");
        ...
        int64_t speed = 100;

Signed-off-by: Fam Zheng <famz@redhat.com>