rule.mak: Rewrite unnest-vars
The macro unnest-vars is the most important, complicated and hard to
trace magic in QEMU's build system.

Rewrite it in a (hopefully) clearer way, with comments, to make it
easier to maintain.

Remove DSO_CFLAGS and module-objs-m that are not used.

An improvement of this version is, per object variables are properly
protected in save-objs and load-objs, before including sub-dir
Makefile.objs, just as nested variables are. So the occasional same
object name from different directory levels won't step on each other's
foot.

Signed-off-by: Fam Zheng <famz@redhat.com>