qapi: Treat nested dict as properties in type definition
This changes the parsing of type definition in qapi schema.

Before:

'data' is a complex dict, that can have nested dictionary parsed as
sub-struct.

After:

'data' is at most 2 levels. The first level's keys are argument names;
The values of the name keys are either argument type names as string, or
property dictionary, containing 'type' and 'default'.

Signed-off-by: Fam Zheng <famz@redhat.com>