qapi: Add @arg property dictionary syntax
parse_args now yields one more item in an iteration - default. It is the
default value for the argument, that is parsed from

    'data': { '@ARG_NAME': { PROP_NAME : PROP_VALUE } }

syntax added into qapi schema.

ARG_NAME is the name of argument. PROP_NAME is the property name of
argument, currently can be 'type', 'optional' or 'default'.

'type' defines the type of the argument, as what we have in:

    'data': { 'ARG_NAME': 'TYPE_NAME' }

'optional' can be true or false.

'default' gives the default value of argument. if 'default' is set,
'optional' is ignored and the qmp handler will always get a value, hence
parse_args yields an False, so that has_xxx parameter is not generated
in the handler prototype. Later, in qapi-commands.py, the value will be
used if user doesn't specify one.

See the coming change on block-commit for example of how qmp commands
can make use of this.

Signed-off-by: Fam Zheng <famz@redhat.com>