qapi: Initialize argument value in generated code if has 'default'
Input marshalling code generation is modified.

If an argument has default, initialize its value in generated code to
that default. If this argument is specified by user input, visit
argument with deallocate visitor, before input visitor, so that the
initial value is deallocated.

This moves initialization of md, mi, vd, vi before input block as
necessary, and moves cleanup of mi aside cleanup of md, for symmetry.

Signed-off-by: Fam Zheng <famz@redhat.com>