virtio-scsi: Reject scsi-cd if data plane enabled
[Forwardporting from qemu-kvm-rhev, with commit message tweaked.]

RHBZ: 1378816

Work has long stalled in a gap between BlockBackend and IOThread
AioContext, which is a trouble for CD change. The assertion failure
which can be simply reproduced by eject-then-change a scsi-cd on an
IOThread enabled virtio-scsi controller, has never been addressed:

[*]: hw/scsi/virtio-scsi.c:246: virtio_scsi_ctx_check: Assertion
     `blk_get_aio_context(d->conf.blk) == s->ctx' failed.

As a band-aid, this partially reverts the effect of ef8875b
(virtio-scsi: Remove op blocker for dataplane, since v2.7). We cannot
revert that commit because it's a huge step backward. The plan is to
renovate the interface between BlockBackend and AioContext, so that the
AioContext can be carried forward.

All in all, scsi-cd on virtio-scsi-dataplane has basically been unusable
unless the scsi-cd never enters an empty state, so, disable it
altogether.  Otherwise it would be much more difficult to avoid
crashing.

Signed-off-by: Fam Zheng <famz@redhat.com>