archive-source: Archive whole git clone
We've now switched from archiving individual files listed by "git
ls-files" to cloning the git tree first then generate from there.

We used ls-files so the target files are note included in the archive,
but it's not necessary now. And including .git also makes it possible
for tests to do git commands.

Signed-off-by: Fam Zheng <famz@redhat.com>