virtio-scsi: Fix unplug crash while guest doing IO
At the time main thread doing unplug releases the aio context, the
iothread may start immediately to poll for more requests on cmd VQ.

We'll fail the virtio_scsi_ctx_check() assertion if the unplugged
device gets a new request. The fix is to skip the scsi device if it has
been unplugged.

Reported-by: Zhenwei Pi <pizhenwei@bytedance.com>
Signed-off-by: Fam Zheng <zhengfeiran@bytedance.com>